+++ include/Fauna/world.h
@@ -21,7 +21,7 @@ class SimulationUnit;
 class WorldConstructor;
 
 // Repeat typedef from hft.h
-typedef std::vector< std::shared_ptr<const Hft> > HftList;
+typedef std::vector<std::shared_ptr<const Hft> > HftList;
 
 namespace Output {
 class Aggregator;
@@ -34,7 +34,7 @@ class World {
   /// Constructor: Read parameters and HFTs from instruction file.
   /**
    * \todo Shall this throw an exception on malformed instruction file? The way
-   * the framework() function in LPJ-GUESS is constructed it odes not work well
+   * the framework() function in LPJ-GUESS is constructed it does not work well
    * to catch an exception. So perhaps Fauna::World::World() should just write
    * to stderr and *then* throw an exception?
    *
@@ -46,6 +46,14 @@ class World {
    */
   World(const std::string instruction_filename);
 
+  /// Constructor: Create deactivated `World` object.
+  /**
+   * Even if the megafauna model should be completely deactivated, there might
+   * be a need to create a "dummy" \ref World instance. An object created with
+   * this constructor will not simulate anything.
+   */
+  World();
+
   /// Default destructor.
   ~World();
 
@@ -53,6 +61,11 @@ class World {
   /**
    * The new simulation unit will be owned by \ref World and released as soon
    * as the \ref Habitat instance is marked as to be deleted from extern.
+   *
+   * If the \ref World class was constructed without an instruction file, this
+   * function will do nothing, but will still throw an error if the parameter
+   * is NULL.
+   *
    * \param habitat Pointer to an instance of a class that implements \ref
    * Habitat, given externally by the vegetation model. The World object will
    * release it after it has been marked as dead (\ref Habitat::kill())
@@ -66,9 +79,13 @@ class World {
    * The global megafauna parameters are public because they might be required
    * by the calling vegetation model to check back against vegetation-specific
    * settings.
-   * \throw std::logic_error If the member variable hasn’t been set.
+   * \throw std::logic_error If the member variable hasn’t been set (if the
+   * \ref World class was created without an instruction file).
    */
   const Parameters& get_params() const;
+
+  /// Whether this \ref World object has been created with an instruction file.
+  const bool is_activated() const { return activated; }
 
   /// Iterate through all simulation units and perform simulation for this day.
   /**
@@ -77,6 +94,10 @@ class World {
    *
    * If a \ref Habitat instance is marked as dead, the corresponding simulation
    * unit will be released from memory.
+   *
+   * If the \ref World class was constructed without an instruction file, this
+   * function will do nothing.
+   *
    * \param date The current simulation day.
    * \param do_herbivores Whether to perform herbivore simulations. If false,
    * only the output data of the habitats are updated.
@@ -88,6 +109,9 @@ class World {
  private:
   /// Get the immutable list of herbivore functional types.
   const HftList& get_hfts();
+
+  /// Whether the whole model is active or not.
+  const bool activated;
 
   /// All simulation instructions from the TOML instruction file.
   /**