+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Added
+- `Fauna::World` can be constructed without an instruction file. [#4]
+
 ### Fixed
 - Correctly plot demo results with `demo_results.Rmd` if output is daily or decadal.
 - Fix “First day before last” error in decadal output. [#3]